Improve DishesController

+++ TaisKohtApi/Controllers/api/DishesController.cs
@@ -1,24 +1,30 @@
 ﻿using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using BusinessLogic.DTO;
 using BusinessLogic.Services;
+using Microsoft.AspNet.Identity;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace TaisKohtApi.Controllers.api
 {
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
     [Produces("application/json")]
     [Route("api/v1/Dishes")]
     public class DishesController : Controller
     {
         private readonly IDishService _dishService;
+        private readonly IRestaurantService _restaurantService;
 
-        public DishesController(IDishService dishService)
+        public DishesController(IDishService dishService, IRestaurantService restaurantService)
         {
             _dishService = dishService;
+            _restaurantService = restaurantService;
         }
 
         /// <summary>
@@ -29,6 +35,7 @@ namespace TaisKohtApi.Controllers.api
         /// <response code="429">Too many requests</response>
         /// <response code="500">Internal error, unable to process request</response>
         // GET: api/v1/Dishes
+        [AllowAnonymous]
         [HttpGet]
         [ProducesResponseType(typeof(List<DishDTO>), 200)]
         [ProducesResponseType(404)]
@@ -47,12 +54,13 @@ namespace TaisKohtApi.Controllers.api
         /// <response code="429">Too many requests</response>
         /// <response code="500">Internal error, unable to process request</response>
         // GET: api/v1/Dishes/Daily
+        [AllowAnonymous]
         [HttpGet("Daily")]
         [ProducesResponseType(typeof(List<DishDTO>), 200)]
         [ProducesResponseType(404)]
         [ProducesResponseType(429)]
         [ProducesResponseType(500)]
-        public IActionResult Daily(bool vegan, bool glutenFree, bool lactoseFree)
+        public IActionResult GetDailyDishes(bool vegan, bool glutenFree, bool lactoseFree)
         {
             var result = _dishService.GetAllDailyDishes(vegan, glutenFree, lactoseFree);
             if (!result.Any())
@@ -71,6 +79,7 @@ namespace TaisKohtApi.Controllers.api
         /// <response code="429">Too many requests</response>
         /// <response code="500">Internal error, unable to process request</response>
         // GET: api/v1/Dishes/search?title=th
+        [AllowAnonymous]
         [HttpGet("Search")]
         [ProducesResponseType(typeof(List<DishDTO>), 200)]
         [ProducesResponseType(404)]
@@ -95,12 +104,13 @@ namespace TaisKohtApi.Controllers.api
         /// <response code="429">Too many requests</response>
         /// <response code="500">Internal error, unable to process request</response>
         // GET: api/v1/Dishes/Pricelimit
+        [AllowAnonymous]
         [HttpGet("Pricelimit")]
         [ProducesResponseType(typeof(List<DishDTO>), 200)]
         [ProducesResponseType(404)]
         [ProducesResponseType(429)]
         [ProducesResponseType(500)]
-        public IActionResult Pricelimit(decimal priceLimit)
+        public IActionResult PriceLimit(decimal priceLimit)
         {
             var result = _dishService.SearchDishByPriceLimit(priceLimit);
             if (!result.Any())
@@ -125,7 +135,7 @@ namespace TaisKohtApi.Controllers.api
         [ProducesResponseType(404)]
         [ProducesResponseType(429)]
         [ProducesResponseType(500)]
-        public IActionResult Top(int amount)
+        public IActionResult GetTopDishes(int amount)
         {
             var result = _dishService.GetTopDishes(amount);
             if (!result.Any())
@@ -145,12 +155,13 @@ namespace TaisKohtApi.Controllers.api
         /// <response code="429">Too many requests</response>
         /// <response code="500">Internal error, unable to process request</response>
         // GET: api/v1/Dishes/5
+        [AllowAnonymous]
         [HttpGet("{id}", Name = "GetDish")]
         [ProducesResponseType(typeof(DishDTO), 200)]
         [ProducesResponseType(404)]
         [ProducesResponseType(429)]
         [ProducesResponseType(500)]
-        public IActionResult Get(int id)
+        public IActionResult GetDish(int id)
         {
             var d = _dishService.GetDishById(id);
             if (d == null) return NotFound();
@@ -191,7 +202,9 @@ namespace TaisKohtApi.Controllers.api
         [ProducesResponseType(500)]
         public IActionResult Post([FromBody]PostDishDTO dishDTO)
         {
-            if (!ModelState.IsValid) return BadRequest();
+            if (!ModelState.IsValid) return BadRequest("Invalid fields provided, please double check the parameters");
+            if (dishDTO.RestaurantId.Equals(null)) return BadRequest("Dish is not related any Restaurant");
+            if (!IsRestaurantUserOrAdmin(dishDTO.RestaurantId)) return BadRequest("New dish can only be added by admin or by restaurant user");
 
             var newDish = _dishService.AddNewDish(dishDTO);
 
@@ -232,7 +245,10 @@ namespace TaisKohtApi.Controllers.api
         [ProducesResponseType(500)]
         public IActionResult Put(int id, [FromBody]PostDishDTO dishDTO)
         {
-            if (!ModelState.IsValid) return BadRequest();
+            if (!ModelState.IsValid) return BadRequest("Invalid fields provided, please double check the parameters");
+            if (dishDTO.RestaurantId.Equals(null)) return BadRequest("Dish is not related any Restaurant");
+            if (!IsRestaurantUserOrAdmin(dishDTO.RestaurantId)) return BadRequest("Dish can only be updated by admin or by restaurant user");
+
             var d = _dishService.GetDishById(id);
 
             if (d == null) return NotFound();
@@ -258,8 +274,17 @@ namespace TaisKohtApi.Controllers.api
         {
             var d = _dishService.GetDishById(id);
             if (d == null) return NotFound();
+            if (!IsRestaurantUserOrAdmin(d.RestaurantId)) return BadRequest("Dish can only be deleted by admin or by restaurant user");
             _dishService.DeleteDish(id);
             return NoContent();
+        }
+
+        private Boolean IsRestaurantUserOrAdmin(int restaurantId)
+        {
+            var users = _restaurantService.GetRestaurantUsersById(restaurantId);
+            var userIds = new ArrayList();
+            users.ForEach(u => userIds.Add(u.UserId));
+            return User.IsInRole("admin") || userIds.Contains(User.Identity.GetUserId());
         }
     }
 }
